Moved the Model namespace into ORM so it can be distributed easier
Also moved ModelAwareTrait to Datasource it I think it makes more sense there
