+++ phpunit.xml.dist
@@ -16,7 +16,6 @@
 		<testsuite name="CakePHP Test Suite">
 			<directory>./tests/TestCase/</directory>
 			<!-- Excludes are required in order to let DatabaseSuite decorate the tests -->
-			<exclude>./tests/TestCase/Model/</exclude>
 			<exclude>./tests/TestCase/Database/</exclude>
 			<exclude>./tests/TestCase/ORM/</exclude>
 		</testsuite>
