+++ src/Console/Shell.php
@@ -18,9 +18,9 @@ use Cake\Console\ConsoleIo;
 use Cake\Console\Exception\ConsoleException;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
+use Cake\Datasource\ModelAwareTrait;
 use Cake\Filesystem\File;
 use Cake\Log\LogTrait;
-use Cake\Model\ModelAwareTrait;
 use Cake\Utility\Inflector;
 use Cake\Utility\MergeVariablesTrait;
 use Cake\Utility\String;
