+++ src/Datasource/ModelAwareTrait.php
@@ -12,9 +12,9 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Model;
+namespace Cake\Datasource;
 
-use Cake\Model\Exception\MissingModelException;
+use Cake\Datasource\Exception\MissingModelException;
 use InvalidArgumentException;
 
 /**
