+++ src/ORM/Behavior/TreeBehavior.php
@@ -12,7 +12,7 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Model\Behavior;
+namespace Cake\ORM\Behavior;
 
 use Cake\Datasource\Exception\RecordNotFoundException;
 use Cake\Event\Event;
