+++ tests/TestCase/Datasource/ModelAwareTraitTest.php
@@ -11,9 +11,9 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Model;
+namespace Cake\Test\TestCase\Datasource;
 
-use Cake\Model\ModelAwareTrait;
+use Cake\Datasource\ModelAwareTrait;
 use Cake\TestSuite\TestCase;
 
 /**
