+++ tests/TestCase/ORM/Behavior/CounterCacheBehaviorTest.php
@@ -12,12 +12,12 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\Model\Behavior;
+namespace Cake\Test\TestCase\ORM\Behavior;
 
 use Cake\Database\Query;
 use Cake\Datasource\ConnectionManager;
 use Cake\Event\Event;
-use Cake\Model\Behavior\CounterCacheBehavior;
+use Cake\ORM\Behavior\CounterCacheBehavior;
 use Cake\ORM\Entity;
 use Cake\ORM\Table;
 use Cake\ORM\TableRegistry;
